requirements.txt
aenum==2.1.2
aiohttp==3.7.4.post0
alembic==1.6.4
aniso8601==9.0.1
argh==0.26.2
asn1crypto==1.4.0
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
Authlib==0.10
azure-common==1.1.27
azure-core==1.14.0
azure-storage-blob==12.8.1
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
boto3==1.17.79
botocore==1.20.79
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-default-group==1.2.2
cmake==3.18.4.post1
cryptography==3.4.7
cvxopt==1.2.0
cycler==0.10.0
decorator==4.4.2
dj-database-url==0.5.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-heroku==0.3.1
dnspython==2.1.0
email-validator==1.1.2
face-recognition-models==0.3.0
fasteners==0.15
Flask==1.1.4
Flask-BabelEx==0.9.4
Flask-Cors==3.0.10
Flask-Executor==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.0
flatten-dict==0.1.0
furl==2.1.0
gevent==21.1.2
greenlet==1.1.0
gunicorn==19.10.0
idna==2.10
image==1.5.33
ipython==7.23.1
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==2.6.0
LunarCalendar==0.0.9
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
meltano==1.74.0
meltano-flask-security==0.1.0
monotonic==1.6
msrest==0.6.21
multidict==5.1.0
mysqlclient==2.0.3
networkx==2.5.1
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.5.2.52
opencv-contrib-python-headless==4.5.2.52
opencv-python==4.5.2.52
orderedmultidict==1.0.1
oscrypto==1.2.1
parso==0.8.2
passlib==1.7.4
pathlib2==2.3.5
pathtools==0.1.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
psutil==5.8.0
psycopg2-binary==2.8.6
pycparser==2.20
pycryptodomex==3.10.1
Pygments==2.9.0
pyhocon==0.3.57
pyhumps==1.2.2
PyJWT==2.1.0
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPika==0.25.2
pystan==2.19.1.1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.14.0
python-editor==1.0.4
python-gitlab==1.15.0
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
simplejson==3.17.2
smtpapi==0.4.7
snowflake-connector-python==2.4.3
snowflake-sqlalchemy==1.2.4
soupsieve==2.2.1
speaklater==1.3
SQLAlchemy==1.4.15
sqlparse==0.3.1
traitlets==5.0.5
urllib3==1.26.4
waitress==2.0.0
watchdog==0.9.0
wcwidth==0.2.5
Werkzeug==0.16.1
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.3
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.4.0
